requirements.txt
pandas
matplotlib
numpy
scikit-learn
tensorflow
tensorflow_hub
opencv-python
pillow
streamlit